调整

+++ src/main/java/com/guest/controller/CheckInController.java
@@ -1,23 +1,41 @@
 package com.guest.controller;
 
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+import java.util.List;
 
-import com.guest.pojo.po.*;
-import com.guest.pojo.vo.Response;
-import com.guest.pojo.vo.ResponseMsg;
-import com.guest.service.*;
-import com.guest.utils.JwtUtill;
-import io.swagger.annotations.*;
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.List;
+import com.guest.pojo.po.BookMsg;
+import com.guest.pojo.po.CheckIn;
+import com.guest.pojo.po.Cost;
+import com.guest.pojo.po.CostType;
+import com.guest.pojo.po.Guest;
+import com.guest.pojo.po.Room;
+import com.guest.pojo.vo.Response;
+import com.guest.pojo.vo.ResponseMsg;
+import com.guest.service.BookMsgService;
+import com.guest.service.CheckInService;
+import com.guest.service.CostService;
+import com.guest.service.CostTypeService;
+import com.guest.service.FrontService;
+import com.guest.service.GuestService;
+import com.guest.service.RoomService;
+import com.guest.utils.JwtUtill;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 
 /**
  * <p>
@@ -30,7 +48,7 @@ import java.util.List;
 @CrossOrigin
 @Transactional
 @RestController
-@Api(tags = {"入住信息"})
+@Api(tags = { "入住信息" })
 public class CheckInController {
 	@Autowired
 	private CheckInService checkInService;
@@ -58,22 +76,16 @@ public class CheckInController {
 	 */
 	@PostMapping("/bookCheckIn")
 	@ApiOperation(value = "有预约的用户入住")
-	@ApiImplicitParams({
-			@ApiImplicitParam(name = "Authorization", value = "填前台管理员的token", required = true),
-			@ApiImplicitParam(name = "bookMsgId", value = "预定信息的id", required = true),
-	})
-	@ApiResponses({
-			@ApiResponse(code = 200, message = "请求成功"),
-			@ApiResponse(code = 40000, message = "请求失败"),
-			@ApiResponse(code = 40104, message = "非法操作, 试图操作不属于自己的数据")
-	})
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", value = "填前台管理员的token", required = true),
+			@ApiImplicitParam(name = "bookMsgId", value = "预定信息的id", required = true), })
+	@ApiResponses({ @ApiResponse(code = 200, message = "请求成功"), @ApiResponse(code = 40000, message = "请求失败"),
+			@ApiResponse(code = 40104, message = "非法操作, 试图操作不属于自己的数据") })
 	public Response bookCheckIn(HttpServletRequest request, int bookMsgId) {
 		String num = (String) request.getAttribute("num");
 		if (frontService.getById(num) != null) {
 			BookMsg bookMsg = bookMsgService.getById(bookMsgId);
-			CheckIn checkIn = new CheckIn(0, bookMsg.getGuestIdCard()
-					, bookMsg.getResultRoom(), bookMsg.getFromTime()
-					, bookMsg.getToTime(), 1);
+			CheckIn checkIn = new CheckIn(0, bookMsg.getGuestIdCard(), bookMsg.getResultRoom(), bookMsg.getFromTime(),
+					bookMsg.getToTime(), 1);
 
 			bookMsg.setState(11);
 			checkIn.setState(1);
@@ -86,7 +98,8 @@ public class CheckInController {
 				Cost cost1 = new Cost(0, costType.getId(), bookMsg.getResultRoom(), 1, 0);
 				costService.saveOrUpdate(cost1);
 
-				int costNum = (int) (bookMsg.getToTime().getTime() - bookMsg.getFromTime().getTime()) / 1000 / 60 / 60 / 24;
+				int costNum = (int) (bookMsg.getToTime().getTime() - bookMsg.getFromTime().getTime()) / 1000 / 60 / 60
+						/ 24;
 				costType = costTypes2.get(0);
 				cost1 = new Cost(0, costType.getId(), bookMsg.getResultRoom(), costNum, 1);
 				costService.saveOrUpdate(cost1);
@@ -109,28 +122,24 @@ public class CheckInController {
 	 */
 	@PostMapping("/checkIn")
 	@ApiOperation(value = "没预约的用户入住")
-	@ApiImplicitParams({
-			@ApiImplicitParam(name = "Authorization", value = "填前台管理员的token", required = true),
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", value = "填前台管理员的token", required = true),
 			@ApiImplicitParam(name = "fromTime", value = "预计入住时间", required = true),
 			@ApiImplicitParam(name = "toTime", value = "预计退房时间", required = true),
 			@ApiImplicitParam(name = "contact", value = "客户的联系方式", required = true),
 			@ApiImplicitParam(name = "idCard", value = "客户的身份证号", required = true),
 			@ApiImplicitParam(name = "name", value = "客户的姓名", required = true),
-			@ApiImplicitParam(name = "roomId", value = "房间id", required = true)
-	})
-	@ApiResponses({
-			@ApiResponse(code = 200, message = "请求成功"),
-			@ApiResponse(code = 40000, message = "请求失败"),
-			@ApiResponse(code = 40104, message = "非法操作, 试图操作不属于自己的数据")
-	})
-	public Response checkIn(HttpServletRequest request, Long fromTime, Long toTime, String contact, String idCard, String name, String roomId) {
+			@ApiImplicitParam(name = "roomId", value = "房间id", required = true) })
+	@ApiResponses({ @ApiResponse(code = 200, message = "请求成功"), @ApiResponse(code = 40000, message = "请求失败"),
+			@ApiResponse(code = 40104, message = "非法操作, 试图操作不属于自己的数据") })
+	public Response checkIn(HttpServletRequest request, Long fromTime, Long toTime, String contact, String idCard,
+			String name, String roomId) {
 		String num = (String) request.getAttribute("num");
 		if (frontService.getById(num) != null) {
 			CheckIn checkIn = new CheckIn(0, idCard, roomId, new Timestamp(fromTime), new Timestamp(toTime), 1);
 			Guest guest = new Guest(idCard, name, contact);
 			guestService.saveOrUpdate(guest);
 			Room room = roomService.getById(roomId);
-			int maxNum = room.getMaxNum();
+			int maxNum = (int) room.getMaxNum();
 			int num1 = checkInService.getNum(checkIn.getRoomId());
 			if (maxNum > num1) {
 				if (num1 == 0) {
@@ -168,16 +177,12 @@ public class CheckInController {
 	 */
 	@PostMapping("/checkOut")
 	@ApiOperation(value = "退房")
-	@ApiImplicitParams({
-			@ApiImplicitParam(name = "Authorization", value = "填前台管理员的token", required = true),
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", value = "填前台管理员的token", required = true),
 			@ApiImplicitParam(name = "roomId", value = "房间id", required = true)
 
 	})
-	@ApiResponses({
-			@ApiResponse(code = 200, message = "请求成功"),
-			@ApiResponse(code = 40000, message = "请求失败"),
-			@ApiResponse(code = 40104, message = "非法操作, 试图操作不属于自己的数据")
-	})
+	@ApiResponses({ @ApiResponse(code = 200, message = "请求成功"), @ApiResponse(code = 40000, message = "请求失败"),
+			@ApiResponse(code = 40104, message = "非法操作, 试图操作不属于自己的数据") })
 	public Response checkOut(HttpServletRequest request, String roomId) {
 		String num = (String) request.getAttribute("num");
 		if (frontService.getById(num) != null) {
@@ -200,4 +205,3 @@ public class CheckInController {
 		return new Response(ResponseMsg.ILLEGAL_OPERATION);
 	}
 }
-
